Refactor custom types into their own file
