+++ emulation_system/emulation_system/compose_file_creator/input/hardware_models/hardware_model.py
@@ -14,7 +14,7 @@ from pydantic import (
     validator,
 )
 
-from emulation_system.compose_file_creator.config_file_settings import (
+from emulation_system.compose_file_creator.settings.config_file_settings import (
     EmulationLevels,
     Images,
     SourceRepositories,
