+++ emulation_system/emulation_system/compose_file_creator/input/hardware_models/modules/temperature_module.py
@@ -2,7 +2,7 @@
 from pydantic import Field
 from typing_extensions import Literal
 
-from emulation_system.compose_file_creator.config_file_settings import (
+from emulation_system.compose_file_creator.settings.config_file_settings import (
     EmulationLevels,
     Hardware,
     Images,
