+++ emulation_system/tests/compose_file_creator/input/hardware_models/modules/test_magnetic_module.py
@@ -7,7 +7,7 @@ from pydantic import parse_obj_as
 from emulation_system.compose_file_creator.input.hardware_models import (
     MagneticModuleInputModel,
 )
-from emulation_system.compose_file_creator.config_file_settings import (
+from emulation_system.compose_file_creator.settings.config_file_settings import (
     Hardware,
     EmulationLevels,
     SourceType,
