+++ emulation_system/tests/compose_file_creator/input/hardware_models/modules/test_thermocycler_module.py
@@ -8,7 +8,7 @@ import py
 import pytest
 from pydantic import parse_obj_as
 
-from emulation_system.compose_file_creator.config_file_settings import (
+from emulation_system.compose_file_creator.settings.config_file_settings import (
     EmulationLevels,
     Hardware,
     SourceType,
